Convert HRF to JUNIT format.
On step of the azure pipelines build/test piplines is to publish the
test
results. Besids other, azure pipelines can process the JUNIT format,
luckely
boost is able to store the test results in the JUNIt format. The
following
command will output the test results in the JUNIT format:

bin/mlpack_test --result_code=no --report_level=detailed
--log_level=test_suite
--log_format=JUNIT

however the size of the report file is serval gigabytes, which would be
fine if
azure wouldn't aport the parsing step; even changing the report_level
parameter
dosn't effect the report size, this is clearly an error on the boost
output
routine. So I wrote a simple python script that converts from HRF to the
JUNIt
format.